+++ include/OverlayEngine.hpp
@@ -14,6 +14,7 @@ struct RemoteCursorState {
     unsigned char b;
     unsigned int groupId;
     bool isSelecting;
+    bool isConflictBlocked;
     int selStartX;
     int selStartY;
 };
@@ -34,6 +35,7 @@ public:
 
     void SetScale(float scale) { m_scale = scale; }
     void SetBackend(OverlayBackend backend) { m_backend = backend; }
+    void SetActivePeer(unsigned long long id) { m_activePeerId = id; }
 
 private:
     bool m_active;
@@ -42,6 +44,7 @@ private:
     unsigned char m_colorB = 0;
     float m_scale = 1.0f;
     OverlayBackend m_backend = OverlayBackend::GDI;
+    unsigned long long m_activePeerId = 0;
 
 #ifdef _WIN32
     void* m_hwnd;